Applied requested changes
- Changed file type to .jsx
- Removed Navbar and Dashboard because changes were already merged into AppBarDrawer and Dashboard.js
