+++ client/src/pages/presentational/Friend.jsx
